feat: updateTutorInfo and getAllTutorInfo routes and controllers

+++ server/models/UpcomingSession.ts
@@ -27,7 +27,7 @@ module.exports = (sequelize:any, DataTypes:any) => {
       allowNull: false,
     },
     cost: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.REAL,
       allowNull: false,
     },
     sessionContext: {
